pyproject.toml
[project]
name = "project-av"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "capstone>=5.0.3",
    "langchain>=0.3.18",
    "langchain-community>=0.3.17",
    "langchain-core>=0.3.35",
    "langchain-groq>=0.2.4",
    "langchain-text-splitters>=0.3.6",
    "pefile>=2024.8.26",
    "pyelftools>=0.31",
    "pyside6>=6.8.2.1",
    "python-dotenv>=1.0.1",
    "python-magic>=0.4.27",
    "qemu-qmp>=0.0.3",
    "requests>=2.32.3",
    "yara-python>=4.5.1",
]
